+++ src/components/ModalContent/ModalContent.test.js
@@ -226,28 +226,33 @@ const loan = (actionName) => {
 const renderModalContent = (props) => render(<ModalContent {...props} />);
 
 describe('Modal Content', () => {
+  const commonProps = {
+    isLoading: false,
+    declarationInProgress: false,
+    resources: {},
+    mutator,
+    loanAction: 'declareLost',
+    itemRequestCount: 0,
+    activeRecord: {},
+    onClose: jest.fn(),
+    handleError: jest.fn(),
+    toggleButton: jest.fn(),
+    validateAction: jest.fn(),
+    okapi: {
+      url: 'https://folio-testing-okapi.dev.folio.org',
+      tenant: 'diku',
+      okapiReady: true,
+      authFailure: [],
+      bindings: {},
+      currentUser
+    },
+  };
+
   it('Declare Lost Confirm Dialog box', async () => {
     const actionName = 'recallrequested';
     const props = {
-      isLoading: false,
-      resources: {},
-      mutator,
+      ...commonProps,
       loan: loan(actionName),
-      loanAction: 'declareLost',
-      itemRequestCount: 0,
-      activeRecord: {},
-      onClose: jest.fn(),
-      handleError: jest.fn(),
-      disableButton: jest.fn(),
-      validateAction: jest.fn(),
-      okapi: {
-        url: 'https://folio-testing-okapi.dev.folio.org',
-        tenant: 'diku',
-        okapiReady: true,
-        authFailure: [],
-        bindings: {},
-        currentUser
-      },
     };
     renderModalContent(props);
     expect(screen.getByText('ui-users.loans.declareLostDialogBody')).toBeDefined();
@@ -262,25 +267,8 @@ describe('Modal Content', () => {
   it('Cancel Dialog box', async () => {
     const actionName = 'recallrequested';
     const props = {
-      isLoading: false,
-      resources: {},
-      mutator,
+      ...commonProps,
       loan: loan(actionName),
-      loanAction: 'declareLost',
-      itemRequestCount: 0,
-      activeRecord: {},
-      onClose: jest.fn(),
-      handleError: jest.fn(),
-      disableButton: jest.fn(),
-      validateAction: jest.fn(),
-      okapi: {
-        url: 'https://folio-testing-okapi.dev.folio.org',
-        tenant: 'diku',
-        okapiReady: true,
-        authFailure: [],
-        bindings: {},
-        currentUser
-      },
     };
     renderModalContent(props);
     expect(screen.getByText('ui-users.loans.declareLostDialogBody')).toBeDefined();
@@ -295,25 +283,9 @@ describe('Modal Content', () => {
   it('ClaimedReturn confirm dialog box', async () => {
     const actionName = 'recallrequested';
     const props = {
-      isLoading: false,
-      resources: {},
-      mutator,
+      ...commonProps,
       loan: loan(actionName),
       loanAction: 'claimReturned',
-      itemRequestCount: 0,
-      activeRecord: {},
-      onClose: jest.fn(),
-      handleError: jest.fn(),
-      disableButton: jest.fn(),
-      validateAction: jest.fn(),
-      okapi: {
-        url: 'https://folio-testing-okapi.dev.folio.org',
-        tenant: 'diku',
-        okapiReady: true,
-        authFailure: [],
-        bindings: {},
-        currentUser
-      },
     };
     renderModalContent(props);
     expect(screen.getByText('ui-users.loans.claimReturnedDialogBody')).toBeDefined();
@@ -327,25 +299,9 @@ describe('Modal Content', () => {
   it('Claimed return for missing item check', async () => {
     const actionName = 'claimedReturned';
     const props = {
-      isLoading: false,
-      resources: {},
-      mutator,
+      ...commonProps,
       loan: loan(actionName),
       loanAction: 'markAsMissing',
-      itemRequestCount: 0,
-      activeRecord: {},
-      onClose: jest.fn(),
-      handleError: jest.fn(),
-      disableButton: jest.fn(),
-      validateAction: jest.fn(),
-      okapi: {
-        url: 'https://folio-testing-okapi.dev.folio.org',
-        tenant: 'diku',
-        okapiReady: true,
-        authFailure: [],
-        bindings: {},
-        currentUser
-      },
       user: {
         id: '47f7eaea-1a18-4058-907c-62b7d095c61b',
       },
@@ -359,6 +315,7 @@ describe('Modal Content', () => {
 
     fireEvent.click(document.querySelector('[data-test-dialog-confirm-button="true"]'));
   });
+
   it('Claimed return for Lost item check', async () => {
     const record = [
       {
@@ -370,25 +327,9 @@ describe('Modal Content', () => {
       }];
     const actionName = 'claimedReturned';
     const props = {
-      isLoading: false,
+      ...commonProps,
       resources: resources(record),
-      mutator,
       loan: loan(actionName),
-      loanAction: 'declareLost',
-      itemRequestCount: 0,
-      activeRecord: {},
-      onClose: jest.fn(),
-      handleError: jest.fn(),
-      disableButton: jest.fn(),
-      validateAction: jest.fn(),
-      okapi: {
-        url: 'https://folio-testing-okapi.dev.folio.org',
-        tenant: 'diku',
-        okapiReady: true,
-        authFailure: [],
-        bindings: {},
-        currentUser
-      },
     };
     renderModalContent(props);
     expect(screen.getByText('ui-users.loans.declareLostDialogBody')).toBeDefined();
@@ -411,24 +352,9 @@ describe('Modal Content', () => {
       }];
     const actionName = 'claimedReturned';
     const props = {
-      isLoading: false,
+      ...commonProps,
       resources: resources(record),
-      mutator,
       loan: loan(actionName),
-      loanAction: 'declareLost',
-      itemRequestCount: 0,
-      activeRecord: {},
-      onClose: jest.fn(),
-      handleError: jest.fn(),
-      validateAction: jest.fn(),
-      okapi: {
-        url: 'https://folio-testing-okapi.dev.folio.org',
-        tenant: 'diku',
-        okapiReady: true,
-        authFailure: [],
-        bindings: {},
-        currentUser
-      },
     };
     renderModalContent(props);
     expect(screen.getByText('ui-users.loans.declareLostDialogBody')).toBeDefined();
@@ -439,6 +365,7 @@ describe('Modal Content', () => {
 
     fireEvent.click(document.querySelector('[data-test-dialog-confirm-button="true"]'));
   });
+
   it('Checking if the user has closed fine', async () => {
     const record = [{
       loanId: '40f5e9d9-38ac-458e-ade7-7795bd821652',
@@ -449,24 +376,10 @@ describe('Modal Content', () => {
     }];
     const actionName = 'claimedReturned';
     const props = {
-      isLoading: false,
+      ...commonProps,
       resources: resources(record),
       mutator: mutatorSuccess,
       loan: loan(actionName),
-      loanAction: 'declareLost',
-      itemRequestCount: 0,
-      activeRecord: {},
-      onClose: jest.fn(),
-      handleError: jest.fn(),
-      validateAction: jest.fn(),
-      okapi: {
-        url: 'https://folio-testing-okapi.dev.folio.org',
-        tenant: 'diku',
-        okapiReady: true,
-        authFailure: [],
-        bindings: {},
-        currentUser
-      },
     };
     renderModalContent(props);
     expect(screen.getByText('ui-users.loans.declareLostDialogBody')).toBeDefined();